Refactor Firebase config to inline dev mocks and simplify exports
