+++ firebaseConfig.js
@@ -4,14 +4,61 @@ import { getApp, getApps, initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
 import { getFirestore } from 'firebase/firestore';
 
-// DEVモードの場合はモック設定を使用
+// DEVモードの判定
 const isDev = Constants.expoConfig?.extra?.isDev;
+
+// 🎯 条件分岐を外に出して、常にexportできるようにする
+let app, auth, db;
 
 if (isDev) {
   console.log('🎭 DEVモード: モック設定を使用しています');
-  // モック設定をインポートしてエクスポート
-  const mockConfig = require('./firebaseConfig.mock.js');
-  export const { app, auth, db } = mockConfig;
+  
+  // モック用のダミーユーザー
+  const mockUser = {
+    uid: 'mock-user-123',
+    email: 'test@example.com',
+    displayName: 'テストユーザー',
+    photoURL: null,
+    emailVerified: true,
+  };
+
+  // モック認証オブジェクト
+  auth = {
+    currentUser: mockUser,
+    onAuthStateChanged: (callback) => {
+      setTimeout(() => callback(mockUser), 100);
+      return () => {};
+    },
+  };
+
+  // モックFirestoreオブジェクト
+  db = {
+    collection: () => ({
+      doc: () => ({
+        get: () => Promise.resolve({
+          exists: true,
+          data: () => ({ name: 'テストデータ' }),
+        }),
+        set: () => Promise.resolve(),
+        update: () => Promise.resolve(),
+        delete: () => Promise.resolve(),
+      }),
+      add: () => Promise.resolve({ id: 'mock-doc-id' }),
+      where: () => ({
+        get: () => Promise.resolve({
+          docs: [],
+          forEach: () => {},
+        }),
+      }),
+    }),
+  };
+
+  // モックアプリ
+  app = {
+    name: 'mock-app',
+    options: {},
+  };
+
 } else {
   console.log('🔥 本番モード: 実際のFirebase設定を使用しています');
   
@@ -35,11 +82,12 @@ if (isDev) {
     measurementId: firebaseMeasurementId,
   };
 
-  const app = getApps().length === 0 ? initializeApp(firebaseConfig) : getApp();
-  const auth = getAuth(app);
-  const db = getFirestore(app);
-
-  export { app, auth, db };
+  app = getApps().length === 0 ? initializeApp(firebaseConfig) : getApp();
+  auth = getAuth(app);
+  db = getFirestore(app);
 }
+
+// 🚀 条件分岐の外でexport
+export { app, auth, db };
